requirements.txt
spotipy
flask
flask_cors
mysql-connector-python